+++ app/src/main/java/woowacourse/shopping/ui/fashionlist/ProductClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.productlist
+package woowacourse.shopping.ui.fashionlist
 
 import woowacourse.shopping.domain.product.Product
