+++ app/src/main/java/woowacourse/shopping/ui/fashionlist/ProductListViewType.kt
@@ -1,9 +1,9 @@
-package woowacourse.shopping.ui.productlist
+package woowacourse.shopping.ui.fashionlist
 
 import woowacourse.shopping.domain.product.Product
 
 sealed class ProductListViewType {
-    data class ProductItemType(val product: Product) : ProductListViewType()
+    data class FashionProductItemType(val product: Product) : ProductListViewType()
 
     data object LoadMoreType : ProductListViewType()
 }
